backend/requirements.txt
#this is a file that essentially pre runs everythign you need to install on setup, in essense why its called a requirements file


fastapi
uvicorn[standard]
python-dotenv
supabase
anthropic
scikit-learn
numpy
joblib
httpx